Quarto teste

+++ include/head.h
@@ -1,4 +1,4 @@
 // Copyright "2018" Ariel Batista
 #include<iostream>
 
-unsigned long int contaLinhas(const char* nomeArquivo);
+unsigned int contaLinhas(const char* nomeArquivo);
